+++ app/src/main/java/com/codewithre/simedit/data/remote/response/SavingResponse.kt
@@ -1,6 +1,7 @@
 package com.codewithre.simedit.data.remote.response
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigInteger
 
 data class SavingResponse(
 
@@ -29,7 +30,7 @@ data class SavingItem(
 	val title: String? = null,
 
 	@field:SerializedName("target")
-	val target: Int? = null,
+	val target: BigInteger? = null,
 
 	@field:SerializedName("persentase")
 	val persentase: Int? = null,
@@ -47,7 +48,7 @@ data class SavingItem(
 	val id: Int? = null,
 
 	@field:SerializedName("terkumpul")
-	val terkumpul: Int? = null,
+	val terkumpul: BigInteger? = null,
 
 	@field:SerializedName("email")
 	val email: String? = null,
